Create README.md
Project includes a json file and we are asked to perform a few tasks in data wrangling and importation.
Biggest lesson learned here is to make sure missing values are NaN's and not empty strings. I spent more time that I would have
liked trying to use fillna on values that were empty strings and not NaN's.